[441] WIP; include checkbox for user filter to filter by user type
