+++ app/components/filters/style.scss
@@ -1,4 +1,4 @@
-@import "../../base.scss";
+@import "../base.scss";
 @import "~@ministryofjustice/frontend/moj/settings/assets";
 @import "~@ministryofjustice/frontend/moj/settings/measurements";
 @import "~@ministryofjustice/frontend/moj/helpers/all";
