+++ app/components/filters/view.html.erb
@@ -63,6 +63,15 @@
                 <%= label_tag "text_search", filter_label, class: "govuk-label govuk-label--s" %>
                 <%= text_field_tag "text_search", (filters[:text_search] if filters), spellcheck: false, class: "govuk-input" %>
               </div>
+
+              <div class="govuk-form-group">
+                <% ["admin", "provider"].each do |user_type, _| %>
+                  <div class="govuk-checkboxes__item">
+                    <%= check_box_tag "user_type[]", user_type, checked?(filters, :user_type, user_type), id: "user_type-#{user_type}", class: "govuk-checkboxes__input" %>
+                    <%= label_tag "user_type-#{user_type}", label_for("user_type", user_type), class: "govuk-label govuk-checkboxes__label" %>
+                  </div>
+                <% end %>
+              </div>
             <% end %>
           </form>
         </div>
