+++ app/components/paginated_filter/view.html.erb
@@ -1,5 +1,5 @@
 <div class='moj-filter-layout'>
-  <%= render Providers::Filters::View.new(filters: filters, filter_label: filter_label, filter_actions: filter_actions) %>
+  <%= render Filters::View.new(filters: filters, filter_label: filter_label, filter_actions: filter_actions) %>
 
   <div class='moj-filter-layout__content'>
     <div class="moj-action-bar">
