+++ app/controllers/support/allocations_controller.rb
@@ -19,7 +19,7 @@ module Support
     end
 
     def filters
-      @filters ||= ProviderFilter.new(params: filter_params).filters
+      @filters ||= Support::Filters::AllocationFilter.new(params: filter_params).filters
     end
 
     def filter_params
